#203 extend interface to allow config input
