+++ setup.py
@@ -25,7 +25,7 @@ setup(
     python_requires=">=3.8",
     entry_points={
         "console_scripts": [
-            "netspresso-train = netspresso_trainer.train:netspresso_train",
+            "netspresso-train = netspresso_trainer.train:train_with_yaml",
         ],
     },
     classifiers=[
