+++ src/netspresso_trainer/__init__.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
-from .trainer_common import parse_args_netspresso, set_arguments, trainer
+from .train import parse_args_netspresso, set_arguments, train_with_yaml, train_with_config
+from .cfg import *
 
 version = (Path(__file__).parent / "VERSION").read_text().strip()
